+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add ANSI color codes to stylish format [#529](https://github.com/nbbrd/heylogs/issues/529)
+
 ## [0.17.1] - 2026-04-08
 
 ### Changed